--- alchemy/src/cloudflare/account-id.ts
@@ -1,12 +1,0 @@
-import { getCloudflareUserInfo, type CloudflareAuthOptions } from "./auth.js";
-
-export type CloudflareAccountId = string & {
-  readonly __brand: "CloudflareAccountId";
-};
-
-export async function CloudflareAccountId(
-  options: CloudflareAuthOptions,
-): Promise<CloudflareAccountId> {
-  const userInfo = await getCloudflareUserInfo(options);
-  return userInfo.accounts[0].id as CloudflareAccountId;
-}
