+++ alchemy/src/cloudflare/auth.ts
@@ -2,199 +2,84 @@ import fs from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
 import type { Secret } from "../secret.js";
-import type { CloudflareAccountId } from "./account-id.js";
-import { handleApiError } from "./api-error.js";
 /**
  * Authentication options for Cloudflare API
  */
-export interface CloudflareAuthOptions {
+export type CloudflareAuthOptions =
+  | CloudflareApiKeyAuthOptions
+  | CloudflareApiTokenAuthOptions;
+
+export type CloudflareApiKeyAuthOptions = {
   /**
    * API Key to use with API Key
    */
-  apiKey?: Secret;
+  apiKey: Secret;
+  /**
+   * Email to use with API Key
+   * If not provided, will attempt to discover from Cloudflare API
+   */
+  email: string;
 
+  /**
+   * API Token to use with API Key
+   */
+  apiToken?: undefined;
+};
+
+export function isCloudflareApiKeyAuthOptions(
+  options: CloudflareAuthOptions | undefined,
+): options is CloudflareApiKeyAuthOptions {
+  return options !== undefined && options.apiKey !== undefined;
+}
+
+export type CloudflareApiTokenAuthOptions = {
   /**
    * API Token to use with API Key
    */
   apiToken?: Secret;
 
   /**
-   * Email to use with API Key
-   * If not provided, will attempt to discover from Cloudflare API
+   * API Key to use with API Token
    */
-  email?: string;
-}
+  apiKey?: undefined;
 
-export interface CloudflareAccount {
-  name: string;
-  id: CloudflareAccountId;
-  type: "standard" | "zero_rating" | "full_control";
-  settings: {
-    enforce_twofactor: boolean;
-    api_access_enabled: null;
-    access_approval_expiry: null;
-    abuse_contact_email: null;
-  };
-  legacy_flags: {
-    enterprise_zone_quota: {
-      maximum: number;
-      current: number;
-      available: number;
-    };
-  };
-}
+  /**
+   * Email to use with API Token
+   */
+  email?: undefined;
+};
 
-export interface CloudflareUserInfo {
-  apiToken?: Secret;
-  apiKey?: Secret;
-  email: string;
-  username: string;
-  accounts: CloudflareAccount[];
-  organizations: CloudflareOrganization[];
-  tokenPermissions: string[] | undefined;
-  first_name: string | null;
-  last_name: string | null;
-  telephone: string | null;
-  country: string | null;
-  zipcode: string | null;
-  two_factor_authentication_enabled: boolean;
-  two_factor_authentication_locked: boolean;
-  has_pro_zones: boolean;
-  has_business_zones: boolean;
-  has_enterprise_zones: boolean;
-  suspended: boolean;
-  betas: string[];
-}
-
-export interface CloudflareOrganization {
-  id: string;
-  name: string;
-  status: string;
-  permissions: string[];
-  roles: string[];
-}
-
-const userInfoCache: Record<string, CloudflareUserInfo> = {};
-
-export async function getCloudflareUserInfo(
-  options: CloudflareAuthOptions,
-): Promise<CloudflareUserInfo> {
-  const cacheKey = JSON.stringify({
-    apiKey: options.apiKey?.unencrypted,
-    apiToken: options.apiToken?.unencrypted,
-    email: options.email,
-  });
-  if (userInfoCache[cacheKey]) {
-    return userInfoCache[cacheKey];
-  }
-  const headers = await getCloudflareAuthHeaders(options);
-  const user = await fetch("https://api.cloudflare.com/client/v4/user", {
-    headers,
-  });
-  if (!user.ok) {
-    if (user.status === 403) {
-      throw new Error(
-        "Cloudflare authentication required. Did you forget to login with `wrangler login` or set CLOUDFLARE_API_TOKEN, CLOUDFLARE_API_KEY, or CLOUDFLARE_EMAIL, or CLOUDFLARE_API_KEY?",
-      );
-    }
-    await handleApiError(user, "getting", "user", "user");
-  }
-  const accounts = await fetch(
-    "https://api.cloudflare.com/client/v4/accounts",
-    {
-      headers,
-    },
-  );
-  const userInfo: CloudflareUserInfo = {
-    ...((await user.json()) as any).result,
-    accounts: ((await accounts.json()) as any).result,
-  };
-  userInfoCache[cacheKey] = userInfo;
-  return userInfo;
+export function isCloudflareApiTokenAuthOptions(
+  options: CloudflareAuthOptions | undefined,
+): options is CloudflareApiTokenAuthOptions {
+  return options !== undefined && options.apiToken !== undefined;
 }
 
 export async function getCloudflareAuthHeaders(
-  options: CloudflareAuthOptions = {},
+  options: CloudflareAuthOptions | undefined,
 ): Promise<Record<string, string>> {
-  // Check for API Token (preferred method)
-  const apiToken =
-    options.apiToken?.unencrypted ?? process.env.CLOUDFLARE_API_TOKEN;
-  if (apiToken) {
+  if (isCloudflareApiKeyAuthOptions(options)) {
+    // Global API Key
     return {
-      Authorization: `Bearer ${apiToken}`,
+      "X-Auth-Key": options.apiKey.unencrypted,
+      "X-Auth-Email": options.email,
+    };
+  } else if (isCloudflareApiTokenAuthOptions(options)) {
+    // API Token
+    return {
+      Authorization: `Bearer ${options.apiToken?.unencrypted}`,
     };
   }
-
-  // Check for API Key
-  const apiKey = options.apiKey?.unencrypted ?? process.env.CLOUDFLARE_API_KEY;
-  if (apiKey) {
-    return {
-      "X-Auth-Key": apiKey,
-      "X-Auth-Email":
-        options.email ??
-        process.env.CLOUDFLARE_EMAIL ??
-        (await getUserEmailFromApiKey(apiKey)),
-    };
-  }
-
+  // Wrangler OAuth Token
   const authConfig = await getRefreshedAuthConfig();
   if (authConfig.oauth_token) {
     return {
       Authorization: `Bearer ${authConfig.oauth_token}`,
     };
   }
-
   throw new Error(
-    "Cloudflare authentication required. Did you forget to login with `wrangler login` or set CLOUDFLARE_API_TOKEN, CLOUDFLARE_API_KEY, or CLOUDFLARE_EMAIL, or CLOUDFLARE_API_KEY?",
+    "Cloudflare authentication required. Did you forget to login with `wrangler login` or set CLOUDFLARE_API_TOKEN, CLOUDFLARE_API_KEY?",
   );
-}
-
-const emailCache: Record<string, string> = {};
-
-async function getUserEmailFromApiKey(apiKey: string): Promise<string> {
-  if (emailCache[apiKey]) {
-    return emailCache[apiKey];
-  }
-
-  try {
-    const baseUrl = "https://api.cloudflare.com/client/v4";
-
-    // Call the /user endpoint to get user information
-    const response = await fetch(`${baseUrl}/user`, {
-      headers: {
-        "Content-Type": "application/json",
-        "X-Auth-Key": apiKey,
-      },
-    });
-
-    if (!response.ok) {
-      throw new Error(
-        `Failed to get user information: ${response.status} ${response.statusText}`,
-      );
-    }
-
-    const data = (await response.json()) as {
-      success: boolean;
-      result: {
-        id: string;
-        email: string;
-        name: string;
-        [key: string]: any;
-      };
-    };
-
-    if (!data.success || !data.result || !data.result.email) {
-      throw new Error("Cloudflare API did not return valid user information");
-    }
-
-    emailCache[apiKey] = data.result.email;
-    return data.result.email;
-  } catch (error) {
-    console.error("Error retrieving email from Cloudflare API:", error);
-    throw new Error(
-      "Failed to automatically discover email for API Key authentication",
-    );
-  }
 }
 
 async function refreshAuthToken(
